Change Tag models not to require subclassing.
