+++ lib/acts_as_taggable_on_mongoid/models/concerns/tag_associations.rb
@@ -1,0 +1,15 @@
+module ActsAsTaggableOnMongoid
+  module Models
+    module Concerns
+      module TagAssociations
+        extend ActiveSupport::Concern
+
+        included do
+          ### ASSOCIATIONS:
+
+          has_many :taggings, dependent: :destroy, class_name: "ActsAsTaggableOnMongoid::Models::Tagging"
+        end
+      end
+    end
+  end
+end
