+++ lib/acts_as_taggable_on_mongoid/models/concerns/tag_validations.rb
@@ -1,0 +1,18 @@
+module ActsAsTaggableOnMongoid
+  module Models
+    module Concerns
+      module TagValidations
+        extend ActiveSupport::Concern
+
+        included do
+          ### VALIDATIONS:
+
+          validates :name, presence: true
+          validates :context, presence: true
+          validates :taggable_type, presence: true
+          validates :name, uniqueness: { scope: %i[context taggable_type] }
+        end
+      end
+    end
+  end
+end
