+++ lib/acts_as_taggable_on_mongoid/models/concerns/tagging_associations.rb
@@ -1,0 +1,15 @@
+module ActsAsTaggableOnMongoid
+  module Models
+    module Concerns
+      module TaggingAssociations
+        extend ActiveSupport::Concern
+
+        included do
+          belongs_to :tag, counter_cache: true, inverse_of: :taggings
+          belongs_to :taggable, polymorphic: true
+          # belongs_to :tagger, { polymorphic: true, optional: true }
+        end
+      end
+    end
+  end
+end
