+++ spec/support/models/01_alt_tag.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
 # A class representing all tags that have ever been set on a model.
-class AltTag < ActsAsTaggableOnMongoid::Models::Tag
-  def self.collection_name
-    "acts_as_taggable_on_mongoid_models_alt_tags"
-  end
+class AltTag
+  include ActsAsTaggableOnMongoid::Models::Concerns::TagModel
 
   ### ASSOCIATIONS:
