+++ spec/support/models/02_other_alt_tag.rb
@@ -2,7 +2,9 @@
 
 # A class representing all tags that have ever been set on a model.
 class OtherAltTag < AltTag
-  def self.collection_name
-    "acts_as_taggable_on_mongoid_models_other_alt_tags"
-  end
+  include ActsAsTaggableOnMongoid::Models::Concerns::TagModel
+
+  ### ASSOCIATIONS:
+
+  has_many :taggings, as: :tag, dependent: :destroy, class_name: "AltTagging"
 end
